Make URL events asynchronous
